convert timeout property from int to double
- this allows us to configure the timeout to 0.6 in case of mobile scrolling.
- this fixes an issue that has laid dormant since the built-in properties' migration.

+++ src/test/java/testPackage/mockedTests/GuiVerificationTests.java
@@ -9,7 +9,7 @@ import org.testng.annotations.Test;
 public class GuiVerificationTests {
     private static final ThreadLocal<SHAFT.GUI.WebDriver> driver = new ThreadLocal<>();
     private final By locator = SHAFT.GUI.Locator.hasTagName("input").build();
-    private int defaultElementIdentificationTimeout;
+    private double defaultElementIdentificationTimeout;
 
     @Test
     public void test_textTrimmed1() {
